# api/doc/api/swagger.yaml
basePath: /
definitions:
  presenter.Dandelion:
    properties:
      id:
        type: string
      image:
        items:
          type: integer
        type: array
      impression:
        type: string
      is_native:
        type: boolean
      landmark:
        type: string
      statement:
        type: string
      type:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is Uta API for PLANE_WEB.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Uta API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - text/plain
      parameters:
      - description: メールアドレス
        in: formData
        name: email
        required: true
        type: string
      - description: パスワード
        in: formData
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "401":
          description: Unauthorized
      summary: ログイン
      tags:
      - Users
  /auth/signup:
    post:
      parameters:
      - description: 苗字
        in: formData
        name: last_name
        required: true
        type: string
      - description: 名前
        in: formData
        name: first_name
        required: true
        type: string
      - description: 住所
        in: formData
        name: address
        required: true
        type: string
      - description: 電話
        in: formData
        name: phone_number
        required: true
        type: string
      - description: メールアドレス
        in: formData
        name: email
        required: true
        type: string
      - description: パスワード
        in: formData
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
      summary: アカウント登録
      tags:
      - Users
  /dandelions:
    get:
      consumes:
      - application/json
      description: タンポポの一覧を取得する。
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/presenter.Dandelion'
            type: array
      summary: タンポポ一覧取得
      tags:
      - Dandelions
    post:
      consumes:
      - application/json
      description: タンポポを登録する。
      parameters:
      - description: 写真
        in: query
        name: image
        required: true
        type: string
      - description: 特徴
        in: query
        name: statement
        required: true
        type: string
      - description: 緯度
        in: query
        name: lat
        required: true
        type: string
      - description: 経度
        in: query
        name: lng
        required: true
        type: string
      - description: 目印
        in: query
        name: landmark
        required: true
        type: string
      - description: 場所
        in: query
        name: type
        required: true
        type: string
      - description: 感想
        in: query
        name: impression
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "401":
          description: Unauthorized
      summary: タンポポ登録
      tags:
      - Dandelions
  /dandelions/{id}:
    get:
      consumes:
      - application/json
      description: タンポポの単体を取得する。
      parameters:
      - description: Dandelion ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/presenter.Dandelion'
      summary: タンポポ単体取得
      tags:
      - Dandelions
  /me:
    get:
      consumes:
      - text/plain
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
      summary: プロフィール取得
      tags:
      - Users
swagger: "2.0"